vibration_theory/requirements.txt
streamlit
numpy
matplotlib
pillow
plotly
scipy
pandas
pyarrow